refactor to tauri_legacy
