+++ src-tauri/Cargo.toml
@@ -24,6 +24,7 @@ tauri-plugin-window-state = "2"
 fs_extra = "1.3.0"
 tauri-plugin-fs = "2.3.0"
 tauri-plugin-cli = "2.2.0"
+once_cell = "1.21.3"
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem
