+++ src-tauri/src/main.rs
@@ -9,6 +9,7 @@ mod graphql;
 mod importer;
 mod known_source_ports;
 mod tauri_helpers;
+mod tauri_legacy;
 
 fn main() {
   let schema = async_graphql::Schema::build(
@@ -20,6 +21,10 @@ fn main() {
   .finish();
 
   tauri::Builder::default()
+    .setup(|app| {
+      crate::tauri_legacy::set_app_handle(app.handle().clone());
+      Ok(())
+    })
     .plugin(tauri_plugin_graphql::init(schema))
     .plugin(tauri_plugin_window_state::Builder::default().build())
     // .plugin(tauri_plugin_fs::init())
